import-bank-statements: commit imports to git
If the bank statement directory contains a .git subdirectory,
import-bank-statements creates a new commit with all the newly created
hledger files on the `import` branch.

The name of the import branch can be configured in the configuration
file $XDG_CONFIG_HOME/bank-statement-parser/import.cfg by adding an
entry `import_branch` in the `[git]` section.