BNP: fix parsing of multi-line bank statements
The extraction of the transaction table was matching against a string
that only appears after the table on the last page. If there is more
than one page in the bank statement, we can look for two lines of
random-looking character strings that seem to appear at the bottom of
each page. Those character strings seem to always appear just after the
table, except on the last page where the old regular expression still
matches.