remove noise in PayPal transactions
Instead of transferring the money from the the paypal balancing account
to the paypal assets account and then from there to an expenses account,
each PayPal transaction now directly transfers the money from the
balancing account to an expenses account.
This means that the number of postings per PayPal transaction has been
reduced from 4–6 to only two.

What is currently not handled is the case where one uses money that is
already in one's PayPal account to pay something. But since I don't use
PayPal for money transfers or as a vendor, this case never happens for
me.