Instantiate AccountMapper and TransactionCleaner with rules file
They do not know about the XDG config dir, so give them only the rules
file instead of letting them compute the path to the rules file
themselves. This will later allow us to instantiate those two classes
with rules files situated in the Git repository that we parse which
results in easier to reproduce parsing results than storing the rules
separate from the parsing results.