Add a config class to each Parser.
This config class contains the definition of the default accounts of the
parser. The config classes also contain a `load` classmethod that will
later be used to load account names from a config file. For now, the
`load` classmethod only copies the default accounts and does not read
from a config file, yet.