Determine Parser's config type from type argument.
When we define a parser as
```
class MyParser(Parser[MyParserConfig]):
    ...
```
we know that its config type is MyParserConfig. Therefore the definition
of config_type in each Parser was kind of redundant. I now removed the
config_type in all parser classes and added a new method
Parser.config_type that determines the config type from the type
argument used when defining the class.