ov_chipkaart: Work around mypy errors for bs4.
It seems that recent versions of BeautifulSoup in Debian come with type
stubs. Unfortunately BeautifulSoup does not seem to be built with type
checking in mind, which means that I had to sprinkle quite a few asserts
and casts into the ov_chipkaart code to guide mypy towards the right
types.