abnamro: Fix parsing of short dates.
Previously, I've simply added the bank statement's year to short dates
to form a complete date. This broke when transactions from December 31
appeared on January's bank statement of the next year, where they were
parsed with a date that was one year in the future.
To fix this problem, I now parse short dates relative to the bank
statement's date.